+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Full-text search for dishes by name and by multiple fields ([#24](https://github.com/kooznitsa/sampo/pull/24)).
 - Restaurant and dish filters in API ([#26](https://github.com/kooznitsa/sampo/pull/26)).
-- Filter tests ([#26](https://github.com/kooznitsa/sampo/pull/26)).
+- Tests for filters ([#26](https://github.com/kooznitsa/sampo/pull/26)).
+- Filter dishes by restaurant nearest stations in API and admin ([#28](https://github.com/kooznitsa/sampo/pull/28)).
 
 ### Fixed